Extracted Inventory item(s) to separate files
