+++ Ultima5Redux/PlayerCharacters/InventoryItem.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Text;
-using Ultima5Redux.Data;
-using Ultima5Redux.DayNightMoon;
+﻿using System.Diagnostics.CodeAnalysis;
+// ReSharper disable UnusedAutoPropertyAccessor.Global
 
 namespace Ultima5Redux.PlayerCharacters
 {
@@ -35,450 +29,18 @@ namespace Ultima5Redux.PlayerCharacters
             }
         }
 
-        public InventoryItem(int quantity, string longName, string shortName, int spriteNum) : this (quantity, longName, shortName, "", spriteNum)
+        protected InventoryItem(int quantity, string longName, string shortName, int spriteNum) : this (quantity, longName, shortName, "", spriteNum)
         {
             
         }
 
-        public InventoryItem(int quantity, string longName, string shortName, string findDescription, int spriteNum)
+        protected InventoryItem(int quantity, string longName, string shortName, string findDescription, int spriteNum)
         {
             this.Quantity = quantity;
             this.LongName = longName;
             this.ShortName = shortName;
             this.SpriteNum = spriteNum;
             this.FindDescription = findDescription;
-        }
-    }
-
-    public class Moonstone : InventoryItem
-    {
-        public MoonPhaseReferences.MoonPhases Phase { get; }
-        private readonly Moongates _moongates;
-
-        public Moonstone(MoonPhaseReferences.MoonPhases phase, string longName, string shortName, string findDescription, Moongates moongates, InventoryReference invRef) 
-            : base(0, longName, shortName, findDescription, MOONSTONE_SPRITE)
-        {
-            Phase = phase;
-            _moongates = moongates;
-            InvRef = invRef;
-        }
-
-        private const int MOONSTONE_SPRITE = 281;
-        
-        // we will hold onto this enum for later when we assign custom sprites
-        //public enum ItemTypeEnum { NewMoon = 0, CrescentWaxing, FirstQuarter, GibbousWaxing, FullMoon, GibbousWaning, LastQuarter, CrescentWaning, NoMoon }
-
-        private static string AddSpacesBeforeCaps(string str)
-        {
-            // filthy method from here: https://stackoverflow.com/questions/272633/add-spaces-before-capital-letters
-            return new string(str.SelectMany((c, i) => i > 0 && char.IsUpper(c) ? new[] { ' ', c } : new[] { c }).ToArray());
-        }
-
-        public int MoongateIndex => (int) Phase;
-        
-        public override string LongName => AddSpacesBeforeCaps(Phase.ToString());
-        public override string ShortName => AddSpacesBeforeCaps(Phase.ToString());
-
-        /// <summary>
-        /// If the moonstone is buried, then it's not in your inventory
-        /// otherwise if it is NOT buried, then it has to be in your inventory
-        /// </summary>
-        public override int Quantity
-        {
-            get => _moongates.IsMoonstoneBuried((int) Phase) ? 0 : 1;
-            // filthy hack - if the _moongates is null, then the base constructor has called it and it doesn't matter at that point
-            set => _moongates?.SetMoonstoneBuried((int) Phase, value <= 0);
-        }
-        
-        public override bool HideQuantity => true;
-    }
-    
-    public class SpecialItem : InventoryItem
-    {
-        public override bool HideQuantity
-        {
-            get
-            {
-                if (ItemType == ItemTypeSpriteEnum.Carpet) return false;
-                return true;
-            }
-        }
-
-        public enum ItemTypeEnum
-        {
-            Carpet = 0x20A, Grapple = 0x209, Spyglass = 0x214, HMSCape = 0x215, PocketWatch = 0, BlackBadge = 0x218,
-            WoodenBox = 0x219, Sextant = 0x216
-        };
-
-        public enum ItemTypeSpriteEnum
-        {
-            Carpet = 170, Grapple = 12, Spyglass = 89, HMSCape = 260, PocketWatch = 232, BlackBadge = 281,
-            WoodenBox = 270, Sextant = 256
-        }
-
-        public ItemTypeSpriteEnum ItemType { get; }
-
-        public static ItemTypeEnum GetItemOffset(ItemTypeSpriteEnum itemTypeSpriteEnum)
-        {
-            ItemTypeEnum itemType = (ItemTypeEnum)Enum.Parse(typeof(ItemTypeEnum), itemTypeSpriteEnum.ToString());
-            return itemType;
-        }
-
-        public SpecialItem(ItemTypeSpriteEnum itemType, int quantity, string longName, string shortName) : 
-            base(quantity, longName, shortName, (int)itemType)
-        {
-            ItemType = itemType;
-        }
-    }
-
-    public class Amulet : Armour
-    {
-        public AmuletEnum AmuletType;
-        public enum AmuletEnum { AmuletTurning = 0x247, SpikeCollar = 0x248, Ankh = 0x249 }
-
-        private const int AMULET_SPRITE = 268;
-     
-        public override bool HideQuantity => false;
-
-        public Amulet(AmuletEnum amuletType, DataOvlReference.Equipment equipment, DataOvlReference dataOvlRef, List<byte> gameStateByteArray)
-            : base(equipment, dataOvlRef, gameStateByteArray, (int)amuletType, AMULET_SPRITE)
-        {
-            AmuletType = amuletType;
-        }
-    }
-
-    public class Ring : Armour
-    {
-        public RingEnum RingType;
-
-        public enum RingEnum { RingInvisibility = 0x244, RingProtection = 0x245, RingRegeneration = 0x246 }
-
-        private const int RING_SPRITE = 266;
-
-        public override bool HideQuantity => false;
-
-        public Ring(RingEnum ringType, DataOvlReference.Equipment equipment, DataOvlReference dataOvlRef, List<byte> gameStateByteArray)
-                 : base(equipment, dataOvlRef, gameStateByteArray, (int)ringType, RING_SPRITE)
-        {
-            RingType = ringType;
-        }
-    }
-
-    public class Helm : Armour
-    {
-        public enum HelmEnum { LeatherHelm = 0x21A, ChainCoif = 0x21B, IronHelm = 0x21C, SpikedHelm = 0x21D }
-
-        public HelmEnum HelmType;
-
-        private const int HELM_SPRITE = 265;
-
-        public override bool HideQuantity => false;
-        public Helm(HelmEnum helmType, DataOvlReference.Equipment equipment, DataOvlReference dataOvlRef, List<byte> gameStateByteArray)
-                        : base(equipment, dataOvlRef, gameStateByteArray, (int)helmType, HELM_SPRITE)
-        {
-            HelmType = helmType;
-        }
-    }
-
-    public class ChestArmour : Armour
-    {
-        public enum ChestArmourEnum { ClothArmour = 0x223, LeatherArmour = 0x224, Ringmail = 0x225, ScaleMail = 0x226 , 
-            ChainMail = 0x227, PlateMail = 0x228 , MysticArmour = 0x229 }
-
-        public ChestArmourEnum ChestArmourType;
-
-        private const int CHEST_ARMOUR_SPRITE = 267;
-
-        //public override int AttackStat { get; }
-
-        //public override int DefendStat { get; }
-
-        public override bool HideQuantity => false;
-
-        public ChestArmour(ChestArmourEnum chestArmourType, DataOvlReference.Equipment equipment, DataOvlReference dataOvlRef, List<byte> gameStateByteArray)
-          : base(equipment, dataOvlRef, gameStateByteArray, (int)chestArmourType, CHEST_ARMOUR_SPRITE)
-        {
-            ChestArmourType = chestArmourType;
-        }
-
-        //public ChestArmour(ChestArmourEnum chestArmourType, int quantity, string longName, string shortName, int attackStat,
-        //    int defendStat, DataOvlReference.EQUIPMENT specificEquipment) : 
-        //    base(quantity, longName, shortName, CHEST_ARMOUR_SPRITE, attackStat, defendStat, specificEquipment)
-        //{
-        //    ChestArmourType = chestArmourType;
-        //}
-    }
-    
-    public class Shield : Armour
-    {
-        public enum ShieldTypeEnum { SmallShield = 0x21E, LargeShield = 0x21F,  SpikedShield = 0x220, MagicShield = 0x221, JewelShield = 0x222 }
-        
-        private const int SHIELD_SPRITE = 262;
- 
-        public Shield(ShieldTypeEnum shieldType, DataOvlReference.Equipment equipment, DataOvlReference dataOvlRef, List<byte> gameStateByteArray)
-            : base(equipment, dataOvlRef, gameStateByteArray, (int)shieldType, SHIELD_SPRITE)
-        {
-            ShieldType = shieldType;
-        }
-
-        public ShieldTypeEnum ShieldType { get; }
-
-
-        public override bool HideQuantity => false;
-    }
-
-    public class Weapon : CombatItem
-    {
-        public enum WeaponTypeEnum {
-            SmallShield = 0x21E, LargeShield = 0x21F, SpikedShield = 0x220, MagicShield = 0x221, JewelShield = 0x222,
-            Dagger = 0x22a, Sling, Club, FlamingOil, MainGauche, Spear, ThrowingAxe, ShortSword, Mace,
-            MorningStar, Bow, Arrows, Crossbow, Quarrels, LongSword, TwoHHammer, TwoHAxe, TwoHSword, Halberd, SwordofChaos,
-            MagicBow, SilverSword, MagicAxe, GlassSword, JeweledSword, MysticSword
-        }
-
-        public enum WeaponTypeSpriteEnum {
-            SmallShield = 262, LargeShield = 262, SpikedShield = 262, MagicShield = 262, JewelShield = 262,
-            Dagger = 261, Sling = 261, Club = 261, FlamingOil = 261, MainGauche = 261,
-            Spear = 261, ThrowingAxe = 261, ShortSword = 261, Mace = 261,
-            MorningStar = 261, Bow = 261, Arrows = 261, Crossbow = 261, Quarrels = 261, 
-            LongSword = 261, TwoHHammer = 261, TwoHAxe = 261, TwoHSword = 261, Halberd = 261, 
-            SwordofChaos = 261, MagicBow = 261, SilverSword = 261, MagicAxe = 261, 
-            GlassSword = 261, JeweledSword = 261, MysticSword = 261
-        }
-
-        public WeaponTypeEnum WeaponType { get; }
-
-        public bool IsAmmo { get; }
-
-        public bool IsTwoHanded { get; }
-
-        public bool IsShield { get; }
-
-        public override bool HideQuantity => false;
-
-
-        public Weapon(WeaponTypeEnum weapon, WeaponTypeSpriteEnum sprite, DataOvlReference.Equipment equipment, DataOvlReference dataOvlRef, List<byte> gameStateByteArray)
-            : base (equipment, dataOvlRef, gameStateByteArray, (int)weapon, (int)sprite)
-        {
-            IsAmmo = equipment == DataOvlReference.Equipment.Quarrels || equipment == DataOvlReference.Equipment.Arrows;
-            IsTwoHanded = equipment == DataOvlReference.Equipment.TwoHAxe || equipment == DataOvlReference.Equipment.TwoHSword ||
-                equipment == DataOvlReference.Equipment.TwoHHammer || equipment == DataOvlReference.Equipment.Bow || equipment == DataOvlReference.Equipment.MagicBow ||
-                equipment == DataOvlReference.Equipment.Crossbow || equipment == DataOvlReference.Equipment.Halberd || equipment == DataOvlReference.Equipment.FlamingOil;
-            IsShield = equipment == DataOvlReference.Equipment.SmallShield || equipment == DataOvlReference.Equipment.LargeShield ||
-                equipment == DataOvlReference.Equipment.SpikedShield || equipment == DataOvlReference.Equipment.MagicShield || 
-                equipment == DataOvlReference.Equipment.JewelShield;
-        }
-    }
-
-    public abstract class Armour : CombatItem
-    {
-        public Armour(DataOvlReference.Equipment specificEquipment, DataOvlReference dataOvlRef, List<byte> gameStateByteRef, int nOffset, int nSpriteNum) 
-            : base (specificEquipment, dataOvlRef, gameStateByteRef, nOffset, nSpriteNum)
-        {
-
-        }
-
-        //public Armour(int quantity, string longName, string shortName, int nSpriteNum, int attackStat, int defendStat, DataOvlReference.EQUIPMENT specificEquipment) : 
-        //    base(quantity, longName, shortName, nSpriteNum, attackStat, defendStat, specificEquipment)
-        //{
-        //}
-    }
-
-    public abstract class CombatItem : InventoryItem
-    {
-        public DataOvlReference.Equipment SpecificEquipment;
-
-        public static int GetAttack(DataOvlReference dataOvlRef, int nIndex)
-        {
-            List<byte> attackValueList = dataOvlRef.GetDataChunk(DataOvlReference.DataChunkName.ATTACK_VALUES).GetAsByteList();
-            if (nIndex >= attackValueList.Count)
-            {
-                return 0;
-            }
-            return attackValueList[nIndex];
-        }
-        public static int GetDefense(DataOvlReference dataOvlRef, int nIndex)
-        {
-            List<byte> defenseValueList = dataOvlRef.GetDataChunk(DataOvlReference.DataChunkName.DEFENSE_VALUES).GetAsByteList();
-            if (nIndex >= defenseValueList.Count)
-            {
-                return 0;
-            }
-            return defenseValueList[nIndex];
-        }
-
-        public static int GetRequiredStrength(DataOvlReference dataOvlRef, int nIndex)
-        {
-            List<byte> requiredStrengthValueList = dataOvlRef.GetDataChunk(DataOvlReference.DataChunkName.REQ_STRENGTH_EQUIP).GetAsByteList();
-            if (nIndex >= requiredStrengthValueList.Count)
-            {
-                return 0;
-            }
-            return requiredStrengthValueList[nIndex];
-        }
-
-        public static string GetEquipmentString(DataOvlReference dataOvlRef, int nString)
-        {
-            List<string> equipmentNames = dataOvlRef.GetDataChunk(DataOvlReference.DataChunkName.EQUIP_INDEXES).GetAsStringListFromIndexes();
-            if (nString == 0xFF) return " ";
-            return equipmentNames[nString];
-        }
-
-        public int RequiredStrength { get; }
-        public int AttackStat { get; }
-        public int DefendStat { get; }
-        //public CombatItem(int quantity, string longName, string shortName, int nSpriteNum, int attackStat, int defendStat, DataOvlReference.EQUIPMENT specificEquipment)
-        //    : base(quantity, longName, shortName, nSpriteNum)
-        //{
-        //    attackStat = AttackStat;
-        //    defendStat = DefendStat;
-        //    this.SpecificEquipment = specificEquipment;
-        //}
-        //    ChestArmours.Add(new ChestArmour(chestArmour, gameStateByteArray[(int)chestArmour],
-        //equipmentNames[(int)equipment], equipmentNames[(int)equipment],
-        //       CombatItem.GetAttack(dataOvlRef, (int) equipment),
-        //       CombatItem.GetDefense(dataOvlRef, (int) equipment),equipment));
-        
-            // CombatItem.GetAttack(dataOvlRef, (int)specificEquipment), CombatItem.GetDefense(dataOvlRef, (int)specificEquipment)
-        public CombatItem(DataOvlReference.Equipment specificEquipment, DataOvlReference dataOvlRef, List<byte> gameStateRef, int nOffset, int nSpriteNum) 
-            : base (gameStateRef[nOffset], CombatItem.GetEquipmentString(dataOvlRef, (int)specificEquipment), 
-                  CombatItem.GetEquipmentString(dataOvlRef, (int)specificEquipment), nSpriteNum)
-        {
-            AttackStat = GetAttack(dataOvlRef, (int)specificEquipment);
-            DefendStat = GetDefense(dataOvlRef, (int)specificEquipment);
-            RequiredStrength = GetRequiredStrength(dataOvlRef, (int)specificEquipment);
-            this.SpecificEquipment = specificEquipment;
-        }
-    }
-
-    public class Potion : InventoryItem
-    {
-        public enum PotionColor { Blue = 0x282, Yellow = 0x283, Red = 0x284, Green = 0x285, Orange = 0x286, Purple = 0x287, Black = 0x288, White = 0x289 }
-
-        public PotionColor Color { get; }
-
-        private const int POTION_SPRITE = 259;
-        
-        public override bool HideQuantity { get; } = false;
-
-        public Potion(PotionColor color, int quantity, string longName, string shortName) : base(quantity, longName, shortName, POTION_SPRITE)
-        {
-            Color = color;
-        }
-    }
-
-    public class Spell : InventoryItem
-    {
-        private const int SPRITE_NUM = 260;
-        public override bool HideQuantity { get; } = false;
-
-        public SpellWords SpellIncantation { get; }
-
-        public enum SpellWords
-        {
-            // taking a bit of a risk and just let the subsequent values be assigned since they should be in order
-            In_Lor = 0x24A, Grav_Por, An_Zu, An_Nox, Mani, An_Ylem, An_Sanct, An_Xen_Corp, Rel_Hur, In_Wis, Kal_Xen,
-            In_Xen_Mani, Vas_Lor, Vas_Flam, In_Flam_Grav, In_Nox_Grav, In_Zu_Grav, In_Por, An_Grav, In_Sanct, In_Sanct_Grav, Uus_Por,
-            Des_Por, Wis_Quas, In_Bet_Xen, An_Ex_Por, In_Ex_Por, Vas_Mani, In_Zu, Rel_Tym, In_Vas_Por_Ylem, Quas_An_Wis, In_An,
-            Wis_An_Ylem, An_Xen_Ex, Rel_Xen_Bet, Sanct_Lor, Xen_Corp, In_Quas_Xen, In_Quas_Wis, In_Nox_Hur, In_Quas_Corp, In_Mani_Corp,
-            Kal_Xen_Corp, In_Vas_Grav_Corp, In_Flam_Hur, Vas_Rel_Por, An_Tym
-        }
-
-        public enum SpellWordsCircles
-        {
-            In_Lor = 1, Grav_Por = 1, An_Zu = 1, An_Nox = 1, Mani = 1, An_Ylem = 1, An_Sanct = 2, An_Xen_Corp = 2, Rel_Hur = 2, In_Wis = 2, Kal_Xen = 2,
-            In_Xen_Mani = 2, Vas_Lor = 3, Vas_Flam = 3, In_Flam_Grav = 3, In_Nox_Grav = 3, In_Zu_Grav = 3, In_Por = 3, An_Grav = 4, In_Sanct = 4, In_Sanct_Grav = 4, Uus_Por = 4,
-            Des_Por = 4, Wis_Quas = 4, In_Bet_Xen = 5, An_Ex_Por = 5, In_Ex_Por = 5, Vas_Mani = 5, In_Zu = 5, Rel_Tym = 5, In_Vas_Por_Ylem = 6, Quas_An_Wis = 6, In_An = 6,
-            Wis_An_Ylem = 6, An_Xen_Ex = 6, Rel_Xen_Bet = 6, Sanct_Lor = 7, Xen_Corp = 7, In_Quas_Xen = 7, In_Quas_Wis = 7, In_Nox_Hur = 7, In_Quas_Corp = 7, In_Mani_Corp = 8,
-            Kal_Xen_Corp = 8, In_Vas_Grav_Corp = 8, In_Flam_Hur = 8, Vas_Rel_Por = 8, An_Tym = 8
-        }
-
-        public enum UnpublishedSpells
-        {
-            An_Ylem,
-            In_Xen_Mani,
-            Rel_Xen_Bet,
-            In_Quas_Corp
-        }
-
-        public int MinCircle()
-        {
-            return (int)Enum.Parse(typeof(SpellWordsCircles), SpellIncantation.ToString());
-        }
-
-        public Spell(SpellWords spellWord, int quantity, string longName, string shortName) : base(quantity, longName, shortName, SPRITE_NUM)
-        {
-            SpellIncantation = spellWord;
-        }
-
-        public string GetLiteralTranslation()
-        {
-            string[] spellStrs = SpellIncantation.ToString().Split('_');// .Replace('_', ' ');
-            StringBuilder sb = new StringBuilder();
-            foreach (string str in spellStrs)
-            {
-                sb.Append(Spells.GetLiteralTranslation(str)+" ");
-            }
-            //Debug.Write("SpellString: " + spellStr); ;
-            //return spellStr;
-            return sb.ToString().TrimEnd();//Spells.GetLiteralTranslation(spellStr);
-        }
-
-    }
-    
-    public class Scroll : InventoryItem
-    {
-        public enum ScrollSpells { Vas_Lor = 0 , Rel_Hur, In_Sanct, In_An, In_Quas_Wis, Kal_Xen_Corp, In_Mani_Corp, An_Tym }
-        
-        private const int SCROLL_SPRITE = 260;
-
-        public Spell.SpellWords ScrollSpell { get; }
-
-        public override bool HideQuantity { get; } = false;
-
-        public Scroll(Spell.SpellWords spell, int quantity, string longName, string shortName) : base(quantity, longName, shortName, SCROLL_SPRITE)
-        {
-            this.ScrollSpell = spell;
-        }
-    }
-
-    public class ShadowlordShard : InventoryItem
-    {
-        public enum ShardType { Falsehood, Hatred, Cowardice};
-
-        public ShardType Shard { get; }
-        public string EquipMessage { get; }
-
-        private const int SHARD_SPRITE = 436;
-        public override bool HideQuantity { get; } = true;
-
-        public ShadowlordShard(ShardType shardType, int quantity, string longName, string equipMessage) : base(quantity, longName, longName, SHARD_SPRITE)
-        {
-            Debug.WriteLine("Shard: " + shardType.ToString());
-            Shard = shardType;
-            EquipMessage = equipMessage;
-        }
-    }
-    
-    public class LordBritishArtifact : InventoryItem
-    {
-        public enum ArtifactType { Amulet = 439, Crown = 437, Sceptre = 438};
-
-        public string EquipMessage { get; }
-
-        public bool HasItem()
-        {
-            return Quantity != 0;
-        }
-
-        public override bool HideQuantity { get; } = true;
-
-        public ArtifactType Artifact { get; }
-        
-        public LordBritishArtifact(ArtifactType artifact, int quantity, string longName, string equipMessage) : base(quantity, longName, longName, (int)artifact)
-        {
-            Artifact = artifact;
-            EquipMessage = equipMessage;
         }
     }
 }
