+++ Ultima5Redux/PlayerCharacters/Reagent.cs
@@ -1,35 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using Ultima5Redux.Data;
-
-namespace Ultima5Redux.PlayerCharacters
+﻿namespace Ultima5Redux.PlayerCharacters
 {
-    public class Reagents : InventoryItems<Reagent.ReagentTypeEnum, Reagent>
-    {
-        public Reagents(DataOvlReference dataOvlRef, List<byte> gameStateByteArray) : base(dataOvlRef, gameStateByteArray)
-        {
-            int nIndex = 0;
-            foreach (Reagent.ReagentTypeEnum reagent in Enum.GetValues(typeof(Reagent.ReagentTypeEnum)))
-            {
-                AddReagent(reagent, (DataOvlReference.ReagentStrings)nIndex++);
-            }
-        }
-
-        private void AddReagent(Reagent.ReagentTypeEnum reagentType, DataOvlReference.ReagentStrings reagentStrRef )
-        {
-            Reagent reagent = new Reagent(reagentType,
-             GameStateByteArray[(int)reagentType],
-             DataOvlRef.StringReferences.GetString(reagentStrRef),
-             DataOvlRef.StringReferences.GetString(reagentStrRef));
-            Items[reagentType] = reagent;
-        }
-
-        //public override Dictionary<Potion.PotionColor, Potion> Items { get; } = new Dictionary<Potion.PotionColor, Potion>(8);
-
-        public override Dictionary<Reagent.ReagentTypeEnum, Reagent> Items { get; } = new Dictionary<Reagent.ReagentTypeEnum, Reagent>();
-    }
-
-
     public class Reagent : InventoryItem
     {
         private const int REAGENT_SPRITE = 259;
@@ -51,8 +21,5 @@ namespace Ultima5Redux.PlayerCharacters
         //0x2B1 1 0-99 Mandrake Root
         public enum ReagentTypeEnum { SulfurAsh = 0x2AA , Ginseng = 0x2AB, Garlic = 0x2AC, SpiderSilk = 0x2AD, BloodMoss = 0x2AE, BlackPearl = 0x2AF, 
             NightShade = 0x2B0, MandrakeRoot = 0x2B1 };
-
-        
-
     }
 }
