+++ Ultima5Redux/Ultima5Redux.csproj
@@ -108,7 +108,17 @@
     <Compile Include="DayNightMoon\MoonPhaseReferences.cs" />
     <Compile Include="Maps\TileReference.cs" />
     <Compile Include="Maps\TileReferences.cs" />
+    <Compile Include="PlayerCharacters\Amulet.cs" />
+    <Compile Include="PlayerCharacters\Armour.cs" />
+    <Compile Include="PlayerCharacters\Armours.cs" />
+    <Compile Include="PlayerCharacters\ChestArmour.cs" />
+    <Compile Include="PlayerCharacters\CombatItem.cs" />
+    <Compile Include="PlayerCharacters\Helm.cs" />
     <Compile Include="PlayerCharacters\InventoryReference.cs" />
+    <Compile Include="PlayerCharacters\LordBritishArtifact.cs" />
+    <Compile Include="PlayerCharacters\LordBritishArtifacts.cs" />
+    <Compile Include="PlayerCharacters\Moonstone.cs" />
+    <Compile Include="PlayerCharacters\Moonstones.cs" />
     <Compile Include="PlayerCharacters\PlayerCharacterRecord.cs" />
     <Compile Include="PlayerCharacters\PlayerCharacterRecords.cs" />
     <Compile Include="MapCharacters\MapCharacterAnimationState.cs" />
@@ -123,6 +133,21 @@
     <Compile Include="MapCharacters\MapCharacters.cs" />
     <Compile Include="MapCharacters\NonPlayerCharacterMovement.cs" />
     <Compile Include="MapCharacters\NonPlayerCharacterReference.cs" />
+    <Compile Include="PlayerCharacters\Potion.cs" />
+    <Compile Include="PlayerCharacters\Potions.cs" />
+    <Compile Include="PlayerCharacters\Reagents.cs" />
+    <Compile Include="PlayerCharacters\Ring.cs" />
+    <Compile Include="PlayerCharacters\Scroll.cs" />
+    <Compile Include="PlayerCharacters\Scrolls.cs" />
+    <Compile Include="PlayerCharacters\ShadowlordShard.cs" />
+    <Compile Include="PlayerCharacters\ShadowlordShards.cs" />
+    <Compile Include="PlayerCharacters\Shield.cs" />
+    <Compile Include="PlayerCharacters\SpecialItem.cs" />
+    <Compile Include="PlayerCharacters\SpecialItems.cs" />
+    <Compile Include="PlayerCharacters\Spell.cs" />
+    <Compile Include="PlayerCharacters\Spells.cs" />
+    <Compile Include="PlayerCharacters\Weapon.cs" />
+    <Compile Include="PlayerCharacters\Weapons.cs" />
     <Compile Include="Point2D.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
